move product list to list sub folder
