+++ frontend/components/product/list/ProductList.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import Container from "../../modules/container/Container";
+import Container from "../../../modules/container/Container";
 
 class ProductList extends React.Component {
   componentDidMount() {
