+++ frontend/components/product/list/ProductListContainer.js
@@ -1,6 +1,6 @@
 import { connect } from "react-redux";
-import { fetchProducts } from "../../actions/product_actions";
-import { fetchReviews } from "../../actions/review_actions";
+import { fetchProducts } from "../../../actions/product_actions";
+import { fetchReviews } from "../../../actions/review_actions";
 import ProductList from "./ProductList";
 
 const mapStateToProps = (state) => {
